Rename Folder HOCs to helperComponents
...because it was really ugly.
