+++ src/main/js/components/header/ColumnFilter.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import ColumnFilterPopup from "./ColumnFilterPopup";
 import i18n from "i18next";
 import classNames from "classnames";
-import connectToAmpersand from "../HOCs/connectToAmpersand";
+import connectToAmpersand from "../helperComponents/connectToAmpersand";
 
 @connectToAmpersand
 class ColumnFilter extends React.Component {
