+++ src/main/js/components/helperComponents/connectToAmpersand.js
