+++ src/main/js/components/helperComponents/withAbortableXhrRequests.jsx
