+++ src/main/js/components/media/overlay/FileChangeUpload.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Dropzone from "react-dropzone";
 import request from "superagent";
-import withAbortableXhrRequest from "../../HOCs/withAbortableXhrRequests";
+import withAbortableXhrRequest from "../../helperComponents/withAbortableXhrRequests";
 import apiUrl from "../../../helpers/apiUrl";
 import ActionCreator from "../../../actions/ActionCreator";
 import ProgressBar from "../ProgressBar.jsx";
