+++ src/main/js/components/media/overlay/SingleFileEdit.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import connectToAmpersand from "../../HOCs/connectToAmpersand";
-import withAbortableXhrRequests from "../../HOCs/withAbortableXhrRequests";
+import connectToAmpersand from "../../helperComponents/connectToAmpersand";
+import withAbortableXhrRequests from "../../helperComponents/withAbortableXhrRequests";
 import Dropzone from "react-dropzone";
 import request from "superagent";
 import ActionCreator from "../../../actions/ActionCreator";
